feat: imp of useSearch + small changes
